# openapi.yaml
openapi: 3.0.3
info:
  title: IoT Metrics API
  version: "1.0.0"
servers:
  - url: http://localhost:8080
paths:
  /devices/{device_id}/config:
    post:
      summary: Configure device thresholds
      description: Configures device thresholds, replacing any existing configuration (upsert)
      operationId: configureDevice
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigureDeviceRequest'
      responses:
        '201':
          description: Created
  /devices/{device_id}/metrics:
    post:
      summary: Record device metric
      description: Records a device metric and triggers an alert if it breaches configured thresholds
      operationId: recordMetric
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordMetricRequest'
      responses:
        '201':
          description: Created
  /devices/{device_id}/alerts:
    get:
      summary: Get device alerts
      description: Retrieves recent device alerts
      operationId: getDeviceAlerts
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: string
        - name: timeframe.start
          in: query
          schema:
            type: string
          description: Filter for alerts after this time
        - name: timeframe.end
          in: query
          schema:
            type: string
          description: Filter for alerts before this time
        - name: page.size
          in: query
          schema:
            type: integer
            format: int32
          description: Maximum number of alerts to return
        - name: page.token
          in: query
          schema:
            type: string
          description: Opaque pagination token
      responses:
        '200':
          description: A page of alerts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDeviceAlertsResponse'
components:
  schemas:
    ConfigureDeviceRequest:
      type: object
      required:
        - temperature_threshold
        - battery_threshold
      properties:
        temperature_threshold:
          type: number
          format: float
          description: Temperature threshold for alerts
        battery_threshold:
          type: integer
          format: int32
          description: Battery level threshold for alerts
    RecordMetricRequest:
      type: object
      required:
        - temperature
        - battery
        - timestamp
      properties:
        temperature:
          type: number
          format: float
          description: Measured temperature
        battery:
          type: integer
          format: int32
          description: Measured battery level
        timestamp:
          type: string
          format: date-time
          description: Time of the metric reading
    GetDeviceAlertsResponse:
      type: object
      properties:
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        next_page_token:
          type: string
          description: Token for the next page of results
    Alert:
      type: object
      description: An alert triggered when a metric breaches its threshold
      properties:
        metric:
          type: string
          description: The metric that triggered the alert (e.g. temperature, battery)
        value:
          type: number
          description: The recorded value that breached the threshold
        threshold:
          type: number
          description: The configured threshold value
        timestamp:
          type: string
          format: date-time
          description: When the alert was triggered